requirements.txt
dnspython==2.4.2
requests==2.31.0
tqdm==4.66.1
certifi==2024.2.2 